backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
cryptography==45.0.7
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==5.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   social-auth-app-django
django-cors-headers==4.8.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via
    #   -r requirements.in
    #   djoser
djoser==2.3.3
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
oauthlib==3.3.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==25.0
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-dotenv==1.1.1
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.2
    # via drf-spectacular
redis==6.4.0
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
social-auth-app-django==5.5.1
    # via djoser
social-auth-core==4.7.0
    # via social-auth-app-django
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via django
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via requests
